Update upgrade instructions to include the down
In some scenarios it makes sense to also down the stack as part of this. Specifically when plugins are updated but not the image itself, as the container will still always pull the latest plugins on restart, but if container has no new image, no restart will happen with docker-compose up only.